Fixed some login issues
