+++ Components/chatList.html
@@ -2,7 +2,7 @@
 
 <graffiti-discover 
     v-slot="{ objects: invitations, isInitialPolling }" 
-    :channels="[$graffitiSession.value.actor]"
+    :channels="[this.username]"
     :schema="this.inviteSchema">
     
     <ul id="chatlist">
