+++ Components/chatWindow.html
@@ -1,9 +1,6 @@
 <section id="chat">
-    <graffiti-discover 
-        v-if="this.channel"
-        v-slot="{ objects: invitations, isInitialPolling }" 
-        :channels="[$graffitiSession.value.actor]"
-        :schema="chatSchema">
+    <graffiti-discover v-if="this.channel" v-slot="{ objects: invitations, isInitialPolling }"
+        :channels="[this.username]" :schema="chatSchema">
         <section class="chatWindow" v-for="invite of invitations">
             <h2>
                 {{invite.value.title}}
@@ -26,8 +23,6 @@
     </graffiti-discover>
 
     <h2 v-else>no chat selected</h2>
-
-    
 
     <graffiti-discover v-slot="{ objects: messageObjects, isInitialPolling }" :channels="[this.channel]" :schema="{
                 properties: {
@@ -77,7 +72,7 @@
     <section id="sendbar" v-if="this.channel">
         <!-- TODO: add in scheduler + activity functionality -->
         <section>
-            <router-link :to="`/scheduler`"><button class="imgIcon"><img src="./Icons/Scheduler Icon.svg"></button></router-link>
+            <button class="imgIcon" @click="toggleScheduler()"><img src="./Icons/Scheduler Icon.svg"></button>
             <!-- <button class="imgIcon"><img src="./Icons/Activity Suggester Icon.svg"></button> -->
         </section>
         <form @submit.prevent="sendMessage()">
@@ -86,5 +81,55 @@
                 <input type="submit" :value="sending? 'sending...' : 'send'" />
             </fieldset>
         </form>
+    </section>
+
+    <!-- SCHEDULER STUFF -->
+    <section class="scheduler hiddenUnlessReveal">
+        <section class="scheduler-setup hiddenUnlessReveal reveal">
+            <h2>scheduler setup</h2>
+
+            <form @submit.prevent="setupScheduler()">
+                <fieldset :disabled="creating">
+                    <input type="text" v-model="schedulerTitle" placeholder="name of scheduler" />
+                    <table>
+                        <tbody>
+                            <tr>
+                                <th><label for="startDate">From</label></th>
+                                <th><label for="endDate">To</label></th>
+                                <th><label for="startTime">From</label></th>
+                                <th><label for="endTime">To</label></th>
+                            </tr>
+                            <tr>
+                                <td><input id="startDate" type="date" v-model="startDate" min="minDate" /></td>
+                                <td><input id="endDate" type="date" v-model="endDate" min="minDate" /></td>
+                                <td><select id="startTime" v-model="startTime"></select></td>
+                                <td><select id="endTime" v-model="endTime"></select></td>
+                            </tr>
+                        </tbody>
+                    </table>
+                    <!-- todo: add rows -->
+                    <input type="submit" :value="creating? 'setting up scheduler...' : 'setup scheduler'" />
+                </fieldset>
+            </form>
+        </section>
+
+        <section class="scheduler-grid hiddenUnlessReveal">
+            <h2>{{this.schedulerTitle}}</h2>
+            <div>
+                <span class="empty"></span>
+                <section id="dateLabels">
+                </section>
+            </div>
+            <div>
+                <section id="timeLabels">
+                </section>
+                <section id="grid">
+                </section>
+            </div>
+            <span>
+                <button id="clearButton" @click="clearAll">clear availability</button>
+                <button @click="sendScheduler">send scheduler</button>
+            </span>
+        </section>
     </section>
 </section>
