+++ Components/scheduler.html
@@ -1,4 +1,4 @@
-<section class="scheduler-setup hiddenUnlessReveal reveal">
+<!-- <section class="scheduler-setup hiddenUnlessReveal reveal">
     <h2>scheduler setup</h2>
 
     <form @submit.prevent="setupScheduler()">
@@ -19,9 +19,9 @@
                         <td><select id="endTime" v-model="endTime"></select></td>
                     </tr>
                 </tbody>
-            </table>
+            </table> -->
             <!-- todo: add rows -->
-            <input type="submit" :value="creating? 'setting up scheduler...' : 'setup scheduler'" />
+            <!-- <input type="submit" :value="creating? 'setting up scheduler...' : 'setup scheduler'" />
         </fieldset>
     </form>
 </section>
@@ -43,4 +43,4 @@
         <button id="clearButton" @click="clearAll">clear availability</button>
         <button @click="sendScheduler">send scheduler</button>
     </span>
-</section>
+</section> -->
