+++ Components/userContent.html
@@ -4,6 +4,6 @@
     <Calendar v-else-if="this.view == 'calendar'"></Calendar>
     <section v-else class="sideAndChatLayout">
         <SideBar :username="username" :view="view" :inviteSchema="inviteSchema"></SideBar>
-        <ChatWindow :chatName="chatName" :channel="channel" :inviteSchema="inviteSchema"></ChatWindow>
+        <ChatWindow :username="username" :chatName="chatName" :channel="channel" :inviteSchema="inviteSchema" :profileSchema="profileSchema"></ChatWindow>
     </section>
 </main>
