+++ Scripts/index.js
@@ -6,7 +6,7 @@ import { GraffitiPlugin } from "@graffiti-garden/wrapper-vue";
 import { UserContent } from "./userContent.js";
 import { GeneralContent } from "./generalContent.js";
 import { SetupContent } from "./setupContent.js";
-import { Scheduler } from "./scheduler.js";
+// import { Scheduler } from "./scheduler.js";
 
 const profileSchema = {
   properties: {
@@ -38,7 +38,7 @@ const router = createRouter({
     { path: "/:username/:view/:chatName/:channel", component: UserContent, props: true },
     { path: "/:username/:chatName/:channel", component: UserContent, props: true },
 
-    { path: "/scheduler", component: Scheduler },
+    // { path: "/scheduler", component: Scheduler },
   ],
 });
 
@@ -56,7 +56,7 @@ createApp({
     SetupContent: defineAsyncComponent(SetupContent),
     UserContent: defineAsyncComponent(UserContent),
 
-    Scheduler: defineAsyncComponent(Scheduler),
+    // Scheduler: defineAsyncComponent(Scheduler),
   },
 
   mounted() {
@@ -80,7 +80,7 @@ createApp({
     async setupProfile() {
       const profiles = this.$graffiti.discover(
         // channels
-        [this.$graffitiSession.value.actor],
+        ["ajz-meet-profiles"],
         // schema
         profileSchema
       );
@@ -95,10 +95,14 @@ createApp({
       } else if (profileArray.length == 0) {
         router.push("/profile-setup");
       } else {
-        const profile = profileArray[0];
-        this.username = profile.value.username;
-        if (this.$route.path == "/") {
-          router.push("/" + profile.value.username + "/chats");
+        const profile = profileArray.filter(p => p.actor == this.$graffitiSession.value.actor)[0];
+        if (profile) {
+          this.username = profile.value.username;
+          if (this.$route.path == "/") {
+            router.push("/" + profile.value.username + "/chats");
+          }
+        } else {
+          router.push("/profile-setup");
         }
       }
     },
