+++ Scripts/userContent.js
@@ -68,6 +68,30 @@ export async function UserContent() {
       Calendar: defineAsyncComponent(Calendar),
     },
 
+    // methods: {
+    //   async checkAuth() {
+    //     const profiles = this.$graffiti.discover(
+    //       // channels
+    //       ["ajz-meet-profiles"],
+    //       // schema
+    //       this.profileSchema
+    //     );
+  
+    //     const profileArray = [];
+    //     for await (const { object } of profiles) {
+    //       profileArray.push(object);
+    //     }
+    //     const profile = profileArray.filter(p => p.actor == this.$graffitiSession.value.actor)[0];
+    //     if (profile) {
+    //       if (!profile.value.username == this.$route.path.split("/")[1])
+    //       router.push("/" + profile.value.username + "/chats");
+    //     } else {
+    //       router.push("/profile-setup");
+    //     }
+    //     return false;
+    //   }
+    // },
+
     template: await fetch("./Components/userContent.html").then((r) => r.text()),
   };
 }
